+++ server/src/main.rs
@@ -19,7 +19,9 @@ use crate::config::Config;
 use crate::internal::*;
 use clap::{App, Arg};
 use hyper::service::{make_service_fn, service_fn};
-use petshop_proto::api::{petshop_server::PetshopServer, tfb_server::TfbServer};
+use petshop_proto::api::{
+    example_server::ExampleServer, petshop_server::PetshopServer, tfb_server::TfbServer,
+};
 use tokio::sync::broadcast;
 
 mod api;
@@ -71,33 +73,33 @@ async fn server_run(config: Config) -> Result<()> {
     let (shutdown_tx, shutdown_rx1) = broadcast::channel::<bool>(8);
     let shutdown_rx2 = shutdown_tx.subscribe();
 
+    // Build gRPC health service
+    let (mut health_reporter, health_service) = tonic_health::server::health_reporter();
+
+    // Build API services
+    let api = Api::from_config(&config, shutdown_tx)?;
+
     // FIXME: Additional gRPC services after being defined in proto library
     // must be added/implemented in this crate, and added to the envoy
     // configuration for JSON transcoding
 
-    // Build gRPC health service
-    let (mut health_reporter, health_service) = tonic_health::server::health_reporter();
+    let example_service = MetricsService::wrap(api.metrics(), ExampleServer::new(api.clone()));
+    let example_service = CsrfService::wrap(api.csrf(), example_service);
+    health_reporter.set_serving::<ExampleServer<Api>>().await;
+
+    let petshop_service = MetricsService::wrap(api.metrics(), PetshopServer::new(api.clone()));
+    let petshop_service = CsrfService::wrap(api.csrf(), petshop_service);
     health_reporter.set_serving::<PetshopServer<Api>>().await;
+
+    let tfb_service = MetricsService::wrap(api.metrics(), TfbServer::new(api.clone()));
     health_reporter.set_serving::<TfbServer<Api>>().await;
-
-    // Build API services
-    let petshop = Api::from_config(&config, shutdown_tx)?;
-    let petshop_tfb = petshop.clone();
-    let petshop_internal = petshop.clone();
-    let petshop_metrics = petshop.metrics();
-    let petshop_csrf = petshop.csrf();
-
-    let petshop_service = MetricsService::wrap(petshop_metrics, PetshopServer::new(petshop));
-    let petshop_service = CsrfService::wrap(petshop_csrf, petshop_service);
-
-    let tfb_metrics = petshop_tfb.metrics();
-    let tfb_service = MetricsService::wrap(tfb_metrics, TfbServer::new(petshop_tfb));
 
     // Build and serve tonic api server
     info!("api listening on {}", config.api_addr);
     let api_server = tonic::transport::Server::builder()
         .trace_fn(|_| tracing::info_span!(NAME))
         .add_service(health_service)
+        .add_service(example_service)
         .add_service(petshop_service)
         .add_service(tfb_service)
         .serve_with_shutdown(config.api_addr, shutdown_signal(shutdown_rx1));
@@ -105,7 +107,7 @@ async fn server_run(config: Config) -> Result<()> {
     // Build and serve hyper internal server
     info!("internal listening on {}", config.internal_addr);
     let internal_service = make_service_fn(move |_| {
-        let api = petshop_internal.clone();
+        let api = api.clone();
         async move {
             Ok::<_, Error>(service_fn(move |req| {
                 let api = api.clone();